Add pagination to blog
This change moved /blog to / for th emost part to support pagination. We
still keep the old permalink strucutre though, as we cannot change that
at this time. We also keep the old blog/index.html file for linkings
sake.

Hopefully a future version of Jekyll support pagination with different
root path for the blog